Update CLI tool and version number
